reclassifies zrx with a remark
There is elaborate structure that lets all registered users to participate easily before ICO is finished.
There is also a flaw: See Q11. Maker can cancel the order any time thus effectively ending ICO. As there is no incentive for the maker to do so and otherwise smart contracts are nicely trustless we decided not to fail